+++ src/main.tsx
@@ -3,7 +3,7 @@ import ReactDOM from "react-dom/client"
 import * as S from "@/main.Styles"
 import App from "@/App.tsx"
 import GlobalStyles from "@/styles/GlobalStyles.tsx"
-import { theme } from "@/styles/Theme.ts"
+import { theme } from "@/styles/theme"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { ThemeProvider } from "styled-components"
