+++ src/pages/Signup/Signup.tsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 export default function Signup() {
   return <div>Signup</div>
 }
