+++ src/pages/UserEdit/UserEdit.tsx
@@ -1,5 +1,3 @@
-import React from "react"
-
 export default function UserEdit() {
   return <div>UserEdit</div>
 }
