+++ src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom"
-import { DirectMessage, Home, Login, Profile, Signup, UserEdit } from "@/Pages"
+import { DirectMessage, Home, Login, Profile, Signup, UserEdit } from "@/pages"
 import Navbar from "@/components/Navbar/Navbar"
 
 const YAPRoute = () => {
